Extracts Core Data helpers to new file
